Fix numerical drift issue in FPSControls / Matrix4f::lookAt().
core
- Got rid of boxutils and rectangleutils namespaces.
- Fixed numerical drift issue in FPSControls and Matrix4f::lookAt() by
  renormalizing the up vector.
- Can now divide a Vector2i by a float to get a Vector2f.
- Can now divide a Vector3i by a float to get a Vector3f.
- Can now divide a scalar by a Vector3i to get a reciprocal.

CUDA
- Got rid of arrayutils namespace.
- DeviceArray1D and DeviceQueue now uses size_t correctly.
- DeviceArray1D/2D and DeviceOpaqueArray2D learned to describe themselves
  as a cudaResourceDesc.

+++ CUDA/src/DeviceQueue.h
@@ -19,11 +19,11 @@ class DeviceQueue
 {
 public:
 
-    // initializes a null queue
+    // nitializes a null queue
     DeviceQueue();
 
-    // initializes a queue with a fixed capacity
-    DeviceQueue( int capacity );
+    // nitializes a queue with a fixed capacity
+    DeviceQueue( size_t capacity );
 
     DeviceQueue( const DeviceQueue< T >& copy );
     DeviceQueue( DeviceQueue< T >&& move );
@@ -34,8 +34,8 @@ public:
     bool isNull() const;
     bool notNull() const;
 
-    // returns the maximum number of elements the queue can hold
-    int capacity() const;
+    // Returns the maximum number of elements the queue can hold.
+    size_t capacity() const;
 
     // returns the total size of the queue in bytes, including overhead
     size_t sizeInBytes() const;
@@ -43,7 +43,7 @@ public:
     // resizes the queue:
     //   destroys the existing data
     //   and clears the queue in the process (head and tail pointers are set to 0)
-    void resize( int capacity );
+    void resize( size_t capacity );
 
     // clears the queue by setting head and tail absolute indices back to 0
     // does not actually touch the data
